backend/pyproject.toml
[project]
name = "agents-template"
version = "0.1.0"
description = "FastAPI + LangGraph multi-agent infrastructure with streaming, history, and auto-discovery"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "curl-cffi>=0.14.0",
    "duckpy>=2.1.1",
    "fastapi>=0.135.1",
    "granian>=2.7.2",
    "groq>=0.37.1",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.2.10",
    "langchain-cerebras>=0.8.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-nvidia-ai-endpoints>=1.1.0",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "langgraph-checkpoint-postgres>=3.0.4",
    "markitdown[all]>=0.1.5",
    "orjson>=3.11.7",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "src/api"]

[tool.hatch.build.targets.wheel.sources]
"config" = "config"
"src/api" = "api"
